+++ src/renderer/mod.rs
@@ -30,7 +30,7 @@ impl Renderer {
         }
     }
 
-    pub fn table_object_id(&self, position: &[f64; 2]) -> u32 {
+    pub fn table_object_id(&self, position: &[f64; 2]) -> u128 {
         self.mask_renderer.table_object_id(position)
     }
